+++ tests/test_component_editgrid.py
@@ -25,6 +25,42 @@ class editgridComponentTestCase(ComponentTestCase):
         editGrid = self.builder.components['editGrid']
         self.assertEqual(editGrid.label, 'Edit Grid')
 
+    def test_paths(self):
+        editGrid = self.builder.components['editGrid']
+        # editgrid
+        self.assertEqual(editGrid.builder_path_key, ['editGrid'])
+        self.assertEqual(editGrid.builder_path_label, ['Edit Grid'])
+        self.assertEqual(editGrid.builder_input_path_key, ['editGrid'])
+        self.assertEqual(editGrid.builder_input_path_label, ['Edit Grid'])
+        # editgrid inputs
+        for pos, row_with_components in enumerate(editGrid.rows):
+            for component in row_with_components.input_components.values():
+                if component.key == 'textfield':
+                    self.assertEqual(component.builder_path_key, ['editGrid', 'textField'])
+                    self.assertEqual(component.builder_path_label, ['Edit Grid', 'Text Field'])
+                    self.assertEqual(component.builder_input_path_key, ['editGrid', 'textField'])
+                    self.assertEqual(component.builder_input_path_label, ['Edit Grid', 'Text Field'])
+                if component.key == 'date':
+                    self.assertEqual(component.builder_path_key, ['editGrid', 'datetime'])
+                    self.assertEqual(component.builder_path_label, ['Edit Grid', 'Date'])
+                    self.assertEqual(component.builder_input_path_key, ['editGrid', 'datetime'])
+                    self.assertEqual(component.builder_input_path_label, ['Edit Grid', 'Date'])
+                if component.key == 'textArea':
+                    self.assertEqual(component.builder_path_key, ['editGrid', 'textArea'])
+                    self.assertEqual(component.builder_path_label, ['Edit Grid', 'Text Area'])
+                    self.assertEqual(component.builder_input_path_key, ['editGrid', 'textArea'])
+                    self.assertEqual(component.builder_input_path_label, ['Edit Grid', 'Text Area'])
+                if component.key == 'number':
+                    self.assertEqual(component.builder_path_key, ['editGrid', 'number'])
+                    self.assertEqual(component.builder_path_label, ['Edit Grid', 'Number'])
+                    self.assertEqual(component.builder_input_path_key, ['editGrid', 'number'])
+                    self.assertEqual(component.builder_input_path_label, ['Edit Grid', 'Number'])
+                if component.key == 'checkbox':
+                    self.assertEqual(component.builder_path_key, ['editGrid', 'checkbox'])
+                    self.assertEqual(component.builder_path_label, ['Edit Grid', 'Checkbox'])
+                    self.assertEqual(component.builder_input_path_key, ['editGrid', 'checkbox'])
+                    self.assertEqual(component.builder_input_path_label, ['Edit Grid', 'Checkbox'])
+
     def test_get_row_labels(self):
         builder_editGrid = self.builder.components['editGrid']
         editGrid = self.form.input_components[builder_editGrid.key]
@@ -45,7 +81,6 @@ class editgridComponentTestCase(ComponentTestCase):
         date_values = ['2023-01-02', '2023-05-26', '2023-07-03']
         number_values = [123, 789, 456]
         checkbox_values = [True, False, False]
-        select_values = ['option1', 'option2', '']
         textArea_values = ['The first textarea.\nSuch nice!', '', '']
         for pos, row_with_components in enumerate(editGrid.rows):
             for component in row_with_components.input_components.values():